+ Remove combination components limit. Prepare Archetype for shared component feature. Refactoring.

+++ Core/Caller/CallerImplementation/AttachDetach/AttachDetachExternalCF.cs
@@ -33,7 +33,7 @@ namespace AnotherECS.Core.Caller
         public bool Is { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => true; }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Config<TMemoryAllocatorProvider>(State state, Dependencies* dependencies, ushort callerId)
+        public void Config<TMemoryAllocatorProvider>(State state, Dependencies* dependencies, uint callerId)
            where TMemoryAllocatorProvider : IAllocatorProvider<TAllocator, TAllocator>
         {
             this.state = state;
