+++ Core/Caller/CallerImplementation/Common/NothingCaller.cs
@@ -144,13 +144,13 @@ namespace AnotherECS.Core.Caller
         public void RepairMemoryHandle(ref RepairMemoryContext repairMemoryContext) { }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Config<TMemoryAllocatorProvider>(State state, Dependencies* dependencies, ushort callerId)
+        public void Config<TMemoryAllocatorProvider>(State state, Dependencies* dependencies, uint callerId)
             where TMemoryAllocatorProvider : IAllocatorProvider<TAllocator, TAllocator> { }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Add(ref Dependencies dependencies, uint id, ushort elementId) { }
+        public void Add(ref Dependencies dependencies, uint id, uint elementId) { }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Remove(ref Dependencies dependencies, uint id, ushort elementId) { }
+        public void Remove(ref Dependencies dependencies, uint id, uint elementId) { }
     }
 }
