+++ Core/Caller/CallerImplementation/SetupDefault/DefaultCF.cs
@@ -10,7 +10,7 @@ namespace AnotherECS.Core.Caller
         public State state;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Config<TMemoryAllocatorProvider>(State state, Dependencies* dependencies, ushort callerId)
+        public void Config<TMemoryAllocatorProvider>(State state, Dependencies* dependencies, uint callerId)
             where TMemoryAllocatorProvider : IAllocatorProvider<TAllocator, TAllocator>
         {
             this.state = state;
