+++ Core/Caller/CallerImplementation/Sparse/NonSparseCF.cs
@@ -19,7 +19,7 @@ namespace AnotherECS.Core.Caller
         where TAllocator : unmanaged, IAllocator
         where TDense : unmanaged
     {
-        private ushort _itemId;
+        private uint _itemId;
         private Dependencies* _dependencies;
 
         public bool IsSingleDense { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => false; }
@@ -28,7 +28,7 @@ namespace AnotherECS.Core.Caller
 
         
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Config<TMemoryAllocatorProvider>(State state, Dependencies* dependencies, ushort callerId)
+        public void Config<TMemoryAllocatorProvider>(State state, Dependencies* dependencies, uint callerId)
             where TMemoryAllocatorProvider : IAllocatorProvider<TAllocator, TAllocator>
         { 
             _dependencies = dependencies;
